Made eyes rounder, tweaked head
I made the pupils in the eyes a bit rounder to add some softness to the image.
Aligned the head with the body to be a bit more centred.
tweaked the head to be a different shape, however I am still not happy with it and invite others to have a crack