Updated eyes, beak, and got rid of some stroke lines.
In this edit I have made the eyes smoother and rounder and aligned
them as seen in the famous picture of Tux. In order for this to happen
I had to edit the beak so it looked like a real beak and somewhat like
Tux's. I also got rid of line strokes on the feet and beak. Not sure
where to go from here...